+++ apps/api/src/pkg/cache/interface.ts
@@ -1,5 +1,6 @@
 import type { Context } from "hono";
-
+import { MaybePromise } from "../types/promise";
+import type { CacheNamespaces } from "./namespaces";
 export type Entry<TValue> = {
   value: TValue;
 
@@ -10,21 +11,8 @@ export type Entry<TValue> = {
   staleUntil: number;
 };
 
-export type CacheConfig = {
-  /**
-   * How long an entry should be fresh in milliseconds
-   */
-  fresh: number;
-
-  /**
-   * How long an entry should be stale in milliseconds
-   *
-   * Stale entries are still valid but should be refreshed in the background
-   */
-  stale: number;
-};
-
-export interface Cache<TNamespaces extends Record<string, unknown>> {
+export interface Cache<TNamespaces extends Record<string, unknown> = CacheNamespaces> {
+  tier: string;
   /**
    * Return the cached value
    *
@@ -36,9 +24,7 @@ export interface Cache<TNamespaces extends Record<string, unknown>> {
     c: Context,
     namespace: TName,
     key: string,
-  ) =>
-    | [TNamespaces[TName] | undefined, boolean]
-    | Promise<[TNamespaces[TName] | undefined, boolean]>;
+  ) => MaybePromise<[TNamespaces[TName] | undefined, boolean]>;
 
   /**
    * Sets the value for the given key.
@@ -48,10 +34,10 @@ export interface Cache<TNamespaces extends Record<string, unknown>> {
     namespace: keyof TNamespaces,
     key: string,
     value: TNamespaces[TName],
-  ) => void;
+  ) => MaybePromise<void>;
 
   /**
    * Removes the key from the cache.
    */
-  remove: (c: Context, namespace: keyof TNamespaces, key: string) => void;
+  remove: (c: Context, namespace: keyof TNamespaces, key: string) => MaybePromise<void>;
 }