+++ apps/api/src/pkg/usagelimit/client.ts
@@ -1,4 +1,5 @@
-import { logger, metrics } from "@/pkg/global";
+import { Logger } from "../logging";
+import { Metrics } from "../metrics";
 import {
   LimitRequest,
   LimitResponse,
@@ -10,13 +11,19 @@ import {
 export class DurableUsageLimiter implements UsageLimiter {
   private readonly namespace: DurableObjectNamespace;
   private readonly domain: string;
+  private readonly logger: Logger;
+  private readonly metrics: Metrics;
   constructor(opts: {
     namespace: DurableObjectNamespace;
 
     domain?: string;
+    logger: Logger;
+    metrics: Metrics;
   }) {
     this.namespace = opts.namespace;
     this.domain = opts.domain ?? "unkey.dev";
+    this.logger = opts.logger;
+    this.metrics = opts.metrics;
   }
 
   public async limit(req: LimitRequest): Promise<LimitResponse> {
@@ -32,7 +39,7 @@ export class DurableUsageLimiter implements UsageLimiter {
           body: JSON.stringify(req),
         })
         .catch(async (e) => {
-          logger.warn("calling the usagelimit DO failed, retrying ...", {
+          this.logger.warn("calling the usagelimit DO failed, retrying ...", {
             keyId: req.keyId,
             error: (e as Error).message,
           });
@@ -44,10 +51,10 @@ export class DurableUsageLimiter implements UsageLimiter {
         });
       return limitResponseSchema.parse(await res.json());
     } catch (e) {
-      logger.error("usagelimit failed", { error: e });
+      this.logger.error("usagelimit failed", { keyId: req.keyId, error: (e as Error).message });
       return { valid: false };
     } finally {
-      metrics.emit({
+      this.metrics.emit({
         metric: "metric.usagelimit",
         latency: performance.now() - start,
         keyId: req.keyId,