+++ apps/api/src/routes/legacy_keys_deleteKey.ts
@@ -1,4 +1,3 @@
-import { cache, db } from "@/pkg/global";
 import { App } from "@/pkg/hono/app";
 import { createRoute, z } from "@hono/zod-openapi";
 import { schema } from "@unkey/db";
@@ -34,6 +33,7 @@ export type LegacyKeysDeleteKeyResponse = z.infer<
 export const registerLegacyKeysDelete = (app: App) =>
   app.openapi(route, async (c) => {
     const { keyId } = c.req.param();
+    const { cache, db } = c.get("services");
 
     const data = await cache.withCache(c, "keyById", keyId, async () => {
       const dbRes = await db.query.keys.findFirst({