+++ apps/landing/app/changelog/side-list.tsx
@@ -20,10 +20,10 @@ type SideListProps = {
 
 export function SideList({ logs, className }: SideListProps) {
   return (
-    <div className={cn("lg:flex flex-col hidden", className)}>
-      <ScrollArea className="flex flex-col h-96 pr-20">
+    <div className={cn("lg:flex hidden", className)}>
+      <ScrollArea className="h-96 pr-20">
         {logs?.map((log, _index) => (
-          <Link href={`/changelog/${log.slug}`}>
+          <Link href={`/changelog/#${log.slug}`}>
             <p key={log.slug} className="text-sm text-white text-left mb-6 ">
               {format(new Date(log.frontmatter.date), "MMMM dd, yyyy")}
             </p>