+++ internal/rbac/src/rbac.ts
@@ -1,11 +1,39 @@
 import { type Result, result } from "@unkey/result";
-import type { PermissionQuery } from "./queries";
+import { generateErrorMessage } from "zod-error";
+import { type PermissionQuery, permissionQuerySchema } from "./queries";
 export class RBAC {
   public evaluatePermissions(
     q: PermissionQuery,
     roles: string[],
   ): Result<{ valid: true } | { valid: false; message: string }> {
     return this.evaluateQueryV1(q, roles);
+  }
+  public validateQuery(q: PermissionQuery): Result<{ query: PermissionQuery }> {
+    const validQuery = permissionQuerySchema.safeParse(q);
+    if (!validQuery.success) {
+      const message = generateErrorMessage(validQuery.error.issues, {
+        maxErrors: 1,
+        delimiter: {
+          component: ": ",
+        },
+        path: {
+          enabled: true,
+          type: "objectNotation",
+          label: "",
+        },
+        code: {
+          enabled: true,
+          label: "",
+        },
+        message: {
+          enabled: true,
+          label: "",
+        },
+      });
+      return result.fail({ valid: false, message });
+    }
+
+    return result.success({ query: validQuery.data });
   }
 
   private evaluateQueryV1(