+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @unkey/hono
 
+## 1.1.0
+
+### Minor Changes
+
+- 3d81139: Add environment to context
+
+## 1.0.0
+
+### Major Changes
+
+- 7d94063: Require apiId for verifying keys
+
 ## 0.11.1
 
 ### Patch Changes